Archive my old, unmaintained version of Runkit
If you like runkit and want to make use of it,
checkout https://github.com/zenovich/runkit

They're actually maintaining the project and
you'll have a better experience for it.

If, for some reason, you need the last version
that was here, fear not.

Revision control remembers.
$ git checkout last-sgolemon-version

A backup of this repo will remain at
https://github.com/sgolemon/archive-runkit

Cheers!